services/ai-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.7.0
pydantic-settings==2.3.0
alembic==1.13.2
celery==5.4.0
redis==5.0.4
httpx==0.27.0
openai==1.35.0
google-cloud-vision==3.7.0
mindee==4.14.0
python-multipart==0.0.9
Pillow==10.3.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.30
pika==1.3.2
pytest==8.2.0
pytest-asyncio==0.23.0
